pyproject.toml
[tool.poetry]
name = "ferma-data-extractions"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
poetry = "~1.8.2"
azure-ai-documentintelligence = "^1.0.0b2"
google-cloud-storage = "^2.14.0"
pydantic-settings = "^2.3.0"
google-cloud-pubsub = "^2.21.1"
pandas = "^2.2.1"
SQLAlchemy = "^2.0.28"
sentry-sdk = "^1.40.6"
python-pptx = "^0.6.23"
unstructured = { version = "0.14.4", extras = ["all-docs"] }
structlog = "^24.2.0"
ecs-logging = "^2.1.0"
pymysql = "^1.1.1"
pdf2image = "^1.16.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
